+++ main.go
@@ -10,9 +10,12 @@ import (
 	"github.com/ionos-cloud/ionoscloud-mcp/tools/billing"
 	"github.com/ionos-cloud/ionoscloud-mcp/tools/compute"
 	"github.com/ionos-cloud/ionoscloud-mcp/tools/dns"
+	"github.com/ionos-cloud/ionoscloud-mcp/tools/objectstorage"
 	billSDK "github.com/ionos-cloud/sdk-go-bundle/products/billing/v2"
 	ionos "github.com/ionos-cloud/sdk-go-bundle/products/compute/v2"
 	dnsSDK "github.com/ionos-cloud/sdk-go-bundle/products/dns/v2"
+	objstSDK "github.com/ionos-cloud/sdk-go-bundle/products/objectstorage/v2"
+	objmgmtSDK "github.com/ionos-cloud/sdk-go-bundle/products/objectstoragemanagement/v2"
 	"github.com/ionos-cloud/sdk-go-bundle/shared"
 	"github.com/modelcontextprotocol/go-sdk/mcp"
 )
@@ -36,6 +39,8 @@ func main() {
 	client := ionos.NewAPIClient(cfg)
 	dnsClient := dnsSDK.NewAPIClient(cfg)
 	billingClient := billSDK.NewAPIClient(cfg)
+	objmgmtClient := objmgmtSDK.NewAPIClient(cfg)
+	objstClient := objstSDK.NewAPIClient(cfg)
 
 	server := mcp.NewServer(&mcp.Implementation{
 		Name:    serverName,
@@ -61,6 +66,7 @@ func main() {
 	compute.RegisterAll(server, client)
 	dns.RegisterAll(server, dnsClient)
 	billing.RegisterAll(server, billingClient)
+	objectstorage.RegisterAll(server, objstClient, objmgmtClient)
 
 	if err := server.Run(context.Background(), &mcp.StdioTransport{}); err != nil {
 		log.Fatal(err)